requirements-dev.txt
asgiref==3.3.1
backcall==0.2.0
colorama==0.4.4
decorator==4.4.2
Django==3.1.7
django-bootstrap-form==3.4
django-crispy-forms==1.11.1
django-extensions==3.1.1
django-social-share==2.2.1
djangorestframework==3.12.2
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
Pygments==2.8.0
pytz==2021.1
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
